добавлено Contact

+++ dto/voice.go
@@ -4,7 +4,7 @@ package dto
 type Voice struct {
 	FileId       string `json:"file_id"`
 	FileUniqueId string `json:"file_unique_id"`
-	Duration     int    `json:"duration"`
+	Duration     int64  `json:"duration"`
 	MimeType     string `json:"mime_type,omitempty"`
 	FileSize     string `json:"file_size,omitempty"`
 }
